+++ crates/primitives/src/lib.rs
@@ -47,8 +47,9 @@ pub use bitvec;
 pub use bytecode::*;
 pub use constants::*;
 pub use eip7702::{
-    Authorization, AuthorizationList, Eip7702Bytecode, Eip7702DecodeError, InvalidAuthorization,
-    RecoveredAuthorization, Signature, SignedAuthorization, EIP7702_MAGIC, EIP7702_MAGIC_BYTES,
+    Authorization, AuthorizationList, Eip7702Bytecode, Eip7702DecodeError, PrimitiveSignature,
+    RecoveredAuthority, RecoveredAuthorization, SignedAuthorization, EIP7702_MAGIC,
+    EIP7702_MAGIC_BYTES, EIP7702_MAGIC_HASH,
 };
 pub use env::*;