+++ crates/primitives/src/specification.rs
@@ -29,7 +29,7 @@ pub enum SpecId {
     SHANGHAI = 16,        // Shanghai               17034870 (Timestamp: 1681338455)
     CANCUN = 17,          // Cancun                 19426587 (Timestamp: 1710338135)
     PRAGUE = 18,          // Prague                 TBD
-    PRAGUE_EOF = 19,      // Prague+EOF             TBD
+    OSAKA = 19,           // Prague+EOF             TBD
     #[default]
     LATEST = u8::MAX,
 }
@@ -68,7 +68,8 @@ pub enum SpecId {
     GRANITE = 23,
     HOLOCENE = 24,
     PRAGUE = 25,
-    PRAGUE_EOF = 26,
+    OSAKA = 26,
+    ISTHMUS = 27,
     #[default]
     LATEST = u8::MAX,
 }
@@ -111,7 +112,7 @@ impl From<&str> for SpecId {
             "Shanghai" => Self::SHANGHAI,
             "Cancun" => Self::CANCUN,
             "Prague" => Self::PRAGUE,
-            "PragueEOF" => Self::PRAGUE_EOF,
+            "Osaka" => Self::OSAKA,
             #[cfg(feature = "optimism")]
             "Bedrock" => SpecId::BEDROCK,
             #[cfg(feature = "optimism")]
@@ -126,6 +127,8 @@ impl From<&str> for SpecId {
             "Granite" => SpecId::GRANITE,
             #[cfg(feature = "optimism")]
             "Holocene" => SpecId::HOLOCENE,
+            #[cfg(feature = "optimism")]
+            "Isthmus" => SpecId::ISTHMUS,
             _ => Self::LATEST,
         }
     }
@@ -153,7 +156,7 @@ impl From<SpecId> for &'static str {
             SpecId::SHANGHAI => "Shanghai",
             SpecId::CANCUN => "Cancun",
             SpecId::PRAGUE => "Prague",
-            SpecId::PRAGUE_EOF => "PragueEOF",
+            SpecId::OSAKA => "Osaka",
             #[cfg(feature = "optimism")]
             SpecId::BEDROCK => "Bedrock",
             #[cfg(feature = "optimism")]
@@ -168,6 +171,8 @@ impl From<SpecId> for &'static str {
             SpecId::GRANITE => "Granite",
             #[cfg(feature = "optimism")]
             SpecId::HOLOCENE => "Holocene",
+            #[cfg(feature = "optimism")]
+            SpecId::ISTHMUS => "Isthmus",
             SpecId::LATEST => "Latest",
         }
     }
@@ -214,7 +219,7 @@ spec!(MERGE, MergeSpec);
 spec!(SHANGHAI, ShanghaiSpec);
 spec!(CANCUN, CancunSpec);
 spec!(PRAGUE, PragueSpec);
-spec!(PRAGUE_EOF, PragueEofSpec);
+spec!(OSAKA, OsakaSpec);
 
 spec!(LATEST, LatestSpec);
 
@@ -233,6 +238,8 @@ spec!(FJORD, FjordSpec);
 spec!(GRANITE, GraniteSpec);
 #[cfg(feature = "optimism")]
 spec!(HOLOCENE, HoloceneSpec);
+#[cfg(feature = "optimism")]
+spec!(ISTHMUS, IsthmusSpec);
 
 #[cfg(not(feature = "optimism"))]
 #[macro_export]
@@ -297,8 +304,8 @@ macro_rules! spec_to_generic {
                 use $crate::PragueSpec as SPEC;
                 $e
             }
-            $crate::SpecId::PRAGUE_EOF => {
-                use $crate::PragueEofSpec as SPEC;
+            $crate::SpecId::OSAKA => {
+                use $crate::OsakaSpec as SPEC;
                 $e
             }
         }
@@ -368,8 +375,8 @@ macro_rules! spec_to_generic {
                 use $crate::PragueSpec as SPEC;
                 $e
             }
-            $crate::SpecId::PRAGUE_EOF => {
-                use $crate::PragueEofSpec as SPEC;
+            $crate::SpecId::OSAKA => {
+                use $crate::OsakaSpec as SPEC;
                 $e
             }
             $crate::SpecId::BEDROCK => {
@@ -398,6 +405,10 @@ macro_rules! spec_to_generic {
             }
             $crate::SpecId::HOLOCENE => {
                 use $crate::HoloceneSpec as SPEC;
+                $e
+            }
+            $crate::SpecId::ISTHMUS => {
+                use $crate::IsthmusSpec as SPEC;
                 $e
             }
         }
@@ -445,7 +456,9 @@ mod tests {
         #[cfg(feature = "optimism")]
         spec_to_generic!(HOLOCENE, assert_eq!(SPEC::SPEC_ID, HOLOCENE));
         spec_to_generic!(PRAGUE, assert_eq!(SPEC::SPEC_ID, PRAGUE));
-        spec_to_generic!(PRAGUE_EOF, assert_eq!(SPEC::SPEC_ID, PRAGUE_EOF));
+        spec_to_generic!(OSAKA, assert_eq!(SPEC::SPEC_ID, OSAKA));
+        #[cfg(feature = "optimism")]
+        spec_to_generic!(ISTHMUS, assert_eq!(SPEC::SPEC_ID, ISTHMUS));
         spec_to_generic!(LATEST, assert_eq!(SPEC::SPEC_ID, LATEST));
     }
 }