+++ crates/revm/src/rwasm/mod.rs
@@ -123,7 +123,7 @@ impl<EXT, DB: Database> Rwasm<'_, EXT, DB> {
                 self.clear();
                 e
             })?;
-        let output = self.transact_preverified_inner(initial_gas_spend);
+        let output = self.transact_preverified_inner(initial_gas_spend.initial_gas);
         let output = self.handler.post_execution().end(&mut self.context, output);
         self.clear();
         output
@@ -140,7 +140,7 @@ impl<EXT, DB: Database> Rwasm<'_, EXT, DB> {
         self.handler
             .validation()
             .tx_against_state(&mut self.context)?;
-        Ok(initial_gas_spend)
+        Ok(initial_gas_spend.initial_gas)
     }
 
     /// Transact transaction